Move Hash to eth_api module
